+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 -   Allow chapter navigation to forthcoming chapters if editable.
 -   No error on book domain if it's the book domain.
 -   More robust reference mining.
+-   Fixed missing return statement in paste command.
 
 # 0.5.63 - 2023-11-11